requirements.txt
certifi==2020.12.5
cycler==0.10.0
flow-vis==0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.1
six==1.15.0
tqdm==4.59.0
